mv all governance queries to Query.Governance.
Add stable QueryResult* types for some of them.
