+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -22,6 +22,15 @@ instance Arbitrary QueryResultCommitteeMemberState where
 instance Arbitrary QueryResultCommitteeMembersState where
   arbitrary = genericArbitraryU
 
+instance Arbitrary QueryResultConstitution where
+  arbitrary = genericArbitraryU
+
+instance Arbitrary QueryResultDRepState where
+  arbitrary = genericArbitraryU
+
+instance Arbitrary QueryResultDRepStates where
+  arbitrary = genericArbitraryU
+
 instance Arbitrary QueryPoolStateResult where
   arbitrary = QueryPoolStateResult <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
